+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+* 1.1.0
+
+	+ Some refactor
+		+ Narrowed down construction for Android plugin
+		+ Resolve plugin `Fragment` reference simply using `findFragmentByTag`
+		+ Minor UI tweaks
+
 * 1.0.0
 
 	+ Full project setup